re-factoring app.py to multiple files
