+++ backend-flask/requirements.txt
@@ -1,11 +1,11 @@
 flask
 flask-cors
 
-#opentelemetry-api 
-#opentelemetry-sdk 
-#opentelemetry-exporter-otlp-proto-http 
-#opentelemetry-instrumentation-flask 
-#opentelemetry-instrumentation-requests
+opentelemetry-api 
+opentelemetry-sdk 
+opentelemetry-exporter-otlp-proto-http 
+opentelemetry-instrumentation-flask 
+opentelemetry-instrumentation-requests
 
 aws-xray-sdk
